refact: layout 재구성
